+++ core/src/main/java/liga/medical/medicalmonitoring/core/service/AbstractService.java
@@ -3,13 +3,13 @@ package liga.medical.medicalmonitoring.core.service;
 import java.util.List;
 
 public interface AbstractService<T> {
-    T findById(Integer id);
+    T findById(Long id);
 
     List<T> getAll();
 
     void add(T element);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 
     void update(T element);
 }
